+++ next.config.ts
@@ -22,6 +22,14 @@ const nextConfig: NextConfig = {
   // This tells Next.js to export pages as "folders with an `index.html` file inside"
   // We use this option so we can avoid having the `.html` extension at the end of the page URLs.
   trailingSlash: true,
+
+  webpack: (config) => {
+    config.module.rules.push({
+      test: /\.md$/,
+      type: "asset/source",
+    });
+    return config;
+  },
 };
 
 export default nextConfig;