+++ src/utils/validationSchemas.ts
@@ -267,6 +267,8 @@ export const authValidationSchema = yup.object({
     .string()
     .required("Password is required")
     .min(8, "Password must be at least 8 characters"),
+  student_terms_and_conditions: yup.boolean().transform((value) => value === "on" || value === true).oneOf([true], "Terms and conditions are required"),
+  privacy_policy: yup.boolean().transform((value) => value === "on" || value === true).oneOf([true], "Privacy policy is required"),
 });
 
 export const organisationAuthValidationSchema = yup.object({
@@ -287,6 +289,8 @@ export const organisationAuthValidationSchema = yup.object({
     .string()
     .required("Password is required")
     .min(8, "Password must be at least 8 characters"),
+  organisation_terms_and_conditions: yup.boolean().transform((value) => value === "on" || value === true).oneOf([true], "Terms and conditions are required"),
+  privacy_policy: yup.boolean().transform((value) => value === "on" || value === true).oneOf([true], "Privacy policy is required"),
 });
 
 export const resetPasswordValidationSchema = yup.object({